+++ src/base/ethereum/ethereum.test.ts
@@ -1,4 +1,4 @@
-import { isGoerli, isLocalnet, isMainnet } from "efi/ethereum/ethereum";
+import { isGoerli, isLocalnet, isMainnet } from "base/ethereum/ethereum";
 
 describe("isGoerli", () => {
   test("should return correct boolean", () => {
