+++ src/efi-ui/token/hooks/useTokenPrice.ts
@@ -12,7 +12,7 @@ import {
 } from "efi-ui/curve/pools";
 import { useCurveStablecoinPoolVirtualPrice } from "efi-ui/curve/stablePools";
 import { AddressesJson } from "addresses/addresses";
-import { isMainnet } from "efi/ethereum/ethereum";
+import { isMainnet } from "base/ethereum/ethereum";
 import { getTokenInfo } from "tokenlists/tokenlists";
 import { QueryObserverResult } from "react-query";
 import { Currency, Money } from "ts-money";
