+++ src/efi/crypto/formatChainName.ts
@@ -1,6 +1,6 @@
 import { t } from "ttag";
 
-import { ChainId, ChainNames } from "efi/ethereum/ethereum";
+import { ChainId, ChainNames } from "base/ethereum/ethereum";
 
 export function formatChainName(
   active: boolean,
