+++ src/efi/crypto/getCryptoDecimals.ts
@@ -4,7 +4,7 @@ import {
   CryptoAssetType,
   findTokenContract,
 } from "efi/crypto/CryptoAsset";
-import { NUM_ETH_DECIMALS } from "efi/ethereum/ethereum";
+import { NUM_ETH_DECIMALS } from "base/ethereum/ethereum";
 import { TokenMetadata } from "tokenlists/tokenlists";
 
 export function getCryptoDecimals(
