+++ src/efi/providers/providers.ts
@@ -3,7 +3,7 @@ import { providers } from "ethers";
 import { createAlchemyWeb3 } from "@alch/alchemy-web3";
 
 import { AddressesJson } from "addresses/addresses";
-import { ChainId } from "efi/ethereum/ethereum";
+import { ChainId } from "base/ethereum/ethereum";
 
 const LOCAL_RPC_HOST = "http://127.0.0.1:8545";
 const ALCHEMY_GOERLI_KEY = process.env.NEXT_PUBLIC_GOERLI_ALCHEMY_KEY as string;
