--- src/integrations/efi-etherscan/ETHERSCAN_DOMAIN.ts
@@ -1,7 +1,0 @@
-import { AddressesJson } from "addresses/addresses";
-import { ChainId } from "efi/ethereum/ethereum";
-
-export const ETHERSCAN_DOMAIN =
-  AddressesJson.chainId === ChainId.GOERLI
-    ? "https://goerli.etherscan.io"
-    : "https://etherscan.io";
